Split code into separate helpers
