+++ app/helpers/application_helper.rb
@@ -14,100 +14,6 @@ module ApplicationHelper
     ]
   }
 
-  CIVIL_URLS = [
-    [
-      '/services/end-a-civil-partnership',
-      '/end-civil-partnership',
-      '/end-civil-partnership/grounds-for-ending-a-civil-partnership'
-    ],
-    [
-      '/services/end-a-civil-partnership',
-      '/looking-after-children-divorce',
-      '/money-property-when-relationship-ends'
-    ],
-    [
-      '/services/end-a-civil-partnership',
-      '/end-civil-partnership/file-application',
-      '/end-civil-partnership/file-application',
-      '/get-help-with-court-fees',
-      '/divorce-missing-husband-wife',
-      '/end-civil-partnership/if-your-partner-lacks-mental-capacity',
-      '/divorce-missing-husband-wife'
-    ],
-    [
-      '/services/end-a-civil-partnership',
-      '/end-civil-partnership/apply-for-a-conditional-order',
-      '/find-a-legal-adviser',
-    ],
-    [
-      '/services/end-a-civil-partnership',
-      '/end-civil-partnership/apply-for-a-final-order'
-    ]
-  ]
-
-  DIVORCE_URLS = [
-    [
-      '/services/get-a-divorce', # slight hack to make every step always appear on the main task list page
-      '/divorce',
-      '/divorce/grounds-for-divorce'
-    ],
-    [
-      '/services/get-a-divorce',
-      '/looking-after-children-divorce',
-      '/money-property-when-relationship-ends'
-    ],
-    [
-      '/services/get-a-divorce',
-      '/divorce/file-for-divorce', # this occurs twice
-      '/get-help-with-court-fees',
-      '/divorce-missing-husband-wife', # so does this
-      '/divorce/if-your-husband-or-wife-lacks-mental-capacity'
-    ],
-    [
-      '/services/get-a-divorce',
-      '/divorce/apply-for-decree-nisi',
-      '/find-a-legal-adviser'
-    ],
-    [
-      '/services/get-a-divorce',
-      '/divorce/apply-for-a-decree-absolute'
-    ]
-  ]
-
-  def on_civil_url
-    result = false
-    CIVIL_URLS.each do |url|
-      if url.include? request.path
-        result = true
-      end
-    end
-    result
-  end
-
-  def on_civil_step(step, highlighting = 0)
-    if highlighting and CIVIL_URLS[step - 1].include? request.path # steps start at 1, let's not get confused
-      true
-    end
-  end
-
-  # argh duplication
-
-  def on_divorce_url
-    result = false
-    DIVORCE_URLS.each do |url|
-      if url.include? request.path
-        result = true
-      end
-    end
-    result
-  end
-
-  def on_divorce_step(step, highlighting = 0)
-    if highlighting and DIVORCE_URLS[step - 1].include? request.path # steps start at 1, let's not get confused
-      true
-    end
-  end
-
   def set_tasklist(tasklist)
     session[:tasklist] = tasklist
   end
